requirements.txt
streamlit>=1.31.0
yt-dlp>=2024.10.7
ffmpeg-python>=0.2.0